+++ R/use_lintr.R
@@ -1,6 +1,6 @@
 #' Use lintr in your project
 #'
-#' Create a minimal lintr config file as a starting point for customization
+#' Create a minimal lintr config file as a starting point for customization and add it to the .Rbuildignore
 #'
 #' @param path Path to project root, where a `.lintr` file should be created.
 #' If the `.lintr` file already exists, an error will be thrown.
@@ -42,5 +42,37 @@ use_lintr <- function(path = ".", type = c("tidyverse", "full")) {
     )
   )
   write.dcf(the_config, config_file, width = Inf)
+
+  if (file.exists(file.path(path, "DESCRIPTION"))) {
+    # Some OS can only normalize a path if the associated file or folder exists, so the path needs to be re-normalized
+    tryCatch({
+      pkg_path <- normalizePath(path, mustWork = TRUE, winslash = "/")
+      config_file <- normalizePath(file.path(path, lintr_option("linter_file")), mustWork = TRUE, winslash = "/")
+    }, error = function(e) {
+      stop("No entry could be added to the .Rbuildignore.", call. = FALSE)
+    })
+    # Check if config_file is in package i.e. lintr_option("linter_file") != "../.lintr"
+    if (startsWith(config_file, prefix = pkg_path)) {
+      # Skip a extra character for the leading `/`
+      rel_path <- substring(config_file, first = nchar(pkg_path) + 2L, last = nchar(config_file))
+      ignore_path <- file.path(pkg_path, ".Rbuildignore")
+      if (!file.exists(ignore_path)) file.create(ignore_path)
+      # Follow the same procedure as base R to see if the file is already ignored
+      ignore <- tryCatch({
+        trimws(readLines(ignore_path))
+      }, warning = function(e) {
+        cat(file = ignore_path, "\n", append = TRUE)
+        trimws(readLines(ignore_path))
+      })
+      ignore <- ignore[nzchar(ignore)]
+      already_ignored <-
+        any(vapply(ignore, FUN = grepl, x = rel_path, perl = TRUE, ignore.case = TRUE, FUN.VALUE = logical(1L)))
+      if (!already_ignored) {
+        cat(file = ignore_path, rex::rex(start, rel_path, end), sep = "\n", append = TRUE)
+        message("Adding ", rel_path, " to .Rbuildignore")
+      }
+    }
+  }
+
   invisible(config_file)
 }